Start on a basic skill system.
git-svn-id: https://andors-trail.googlecode.com/svn/trunk@134 08aca716-68be-ccc6-4d58-36f5abd142ac

+++ AndorsTrail/src/com/gpl/rpg/AndorsTrail/view/TraitsInfoView.java
@@ -74,7 +74,7 @@ public final class TraitsInfoView extends TableLayout {
 		}
 		if (traits != null && traits.hasCriticalMultiplierEffect()) {
 			traitsinfo_critical_row2.setVisibility(View.VISIBLE);
-			traitsinfo_criticalhit_multiplier.setText(Integer.toString(traits.criticalMultiplier));
+			traitsinfo_criticalhit_multiplier.setText(Float.toString(traits.criticalMultiplier));
 		} else {
 			traitsinfo_critical_row2.setVisibility(View.GONE);
 		}
